requirements.txt
# Django, the main framework for the api
django>=4.1.4,<4.2.0
djangorestframework>=3.14.0,<3.15.0

djangorestframework-simplejwt>=5.2.2,<5.3.0

# Name generator
Faker>=18.11.1,<18.12.0
hangul-names>=0.4.2,<0.5.0

# Shell
ipython>=8.14.0,<8.15.0

